testing things out
re-wrote my html to fit with the new direction that i plan to go in for my games javascript. Been going back and fourth in codpen trying stuff out but i rememberd that im on a new branch of my code and not my ,main branch so icame back here to update my code.